requirements.txt
annotated-types==0.7.0
anyio==4.6.0
blinker==1.8.2
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
distro==1.9.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
Flask==3.0.3
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
lxml==5.3.0
MarkupSafe==3.0.1
openai==1.51.2
openpyxl==3.1.5
pydantic==2.9.2
pydantic_core==2.23.4
python-docx==1.1.2
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
Werkzeug==3.0.4
zipp==3.20.2

fitz~=0.0.1.dev2